requirements.txt
tensorflow~=2.20.0
numpy~=1.26.4
fastapi~=0.129.0
pillow~=10.2.0
python-multipart
fastapi
uvicorn[standard]